fixes: minor fixes and additions

+++ src/Application/Users/Commands/DeleteUserCommand.cs
@@ -1,4 +1,5 @@
 using Application.Common;
+using Application.Common.Interfaces.Queries;
 using Application.Users.Exceptions;
 using Domain.Models.Users;
 using MediatR;
@@ -15,10 +16,12 @@ public class DeleteUserCommandHandler
     : IRequestHandler<DeleteUserCommand, Result<User, UserException>>
 {
     private readonly UserManager<User> _userManager;
+    private readonly IProjectQueries _projectQueries;
 
-    public DeleteUserCommandHandler(UserManager<User> userManager)
+    public DeleteUserCommandHandler(UserManager<User> userManager, IProjectQueries projectQueries)
     {
         _userManager = userManager;
+        _projectQueries = projectQueries;
     }
 
     public async Task<Result<User, UserException>> Handle(
@@ -32,6 +35,14 @@ public class DeleteUserCommandHandler
             return await Task.FromResult<Result<User, UserException>>(new UserNotFoundException(request.UserId));
         }
 
+        var projectsWithCreator = await _projectQueries.GetAllByCreator(existingUser.Id, cancellationToken);
+        var projectsWithClient = await _projectQueries.GetAllByClient(existingUser.Id, cancellationToken);
+
+        if (projectsWithClient.Count > 0 || projectsWithCreator.Count > 0)
+        {
+            return await Task.FromResult<Result<User, UserException>>(new UserAlreadyUsedInProject(request.UserId));
+        }
+        
         var result = await _userManager.DeleteAsync(existingUser);
         return Result<User, UserException>.FromIdentityResult<User, UserException>(result, existingUser,
                      e => new UserUnknownException(request.UserId, new Exception(e.ToString())));
